+++ src/main/java/lotto/controller/LottoManager.java
@@ -1,6 +1,12 @@
 package lotto.controller;
 
-import lotto.domain.*;
+import lotto.domain.Buyer;
+import lotto.domain.Money;
+import lotto.domain.lottoTicket.LottoNumber;
+import lotto.domain.lottoTicket.WinningLotto;
+import lotto.domain.result.LottoResult;
+import lotto.domain.result.WinningValue;
+import lotto.util.ConvertInput;
 
 import java.util.List;
 import java.util.Map;
