+++ src/main/java/lotto/domain/lottoTicket/LottoFactory.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottoTicket;
 
 import java.util.ArrayList;
 import java.util.Collections;
