+++ src/main/java/lotto/domain/result/LottoResult.java
@@ -1,4 +1,7 @@
-package lotto.domain;
+package lotto.domain.result;
+
+import lotto.domain.lottoTicket.Lotto;
+import lotto.domain.lottoTicket.WinningLotto;
 
 import java.util.*;
