+++ src/test/java/lotto/domain/lottoTicket/LottoFactoryTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottoTicket;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
